# backend/src/Application/Pipelines/Queries/PipelineHeadDto.cs
using KwikDeploy.Application.Common.Mappings;
using KwikDeploy.Domain.Entities;

namespace KwikDeploy.Application.Pipelines.Queries;

public class PipelineHeadDto : IMapFrom<Pipeline>
{
    public int Id { get; init; }

    public string Name { get; init; } = null!;
}